requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
mysql-connector-python
pymysql
SQLAlchemy
python-dotenv
fastapi
uvicorn[standard]
jinja2
python-multipart
evidently
boto3
pyyaml
dill
-e .
setuptools